# Conflicts:
#	src/main/java/com/example/enterprise_internet_applications_project/controllers/FileManagementController.java
#	src/main/java/com/example/enterprise_internet_applications_project/models/Group.java
#	src/main/java/com/example/enterprise_internet_applications_project/repositories/FilesRepository.java
#	src/main/java/com/example/enterprise_internet_applications_project/repositories/PersonRepository.java
#	src/main/java/com/example/enterprise_internet_applications_project/services/FilesService.java

+++ src/main/java/com/example/enterprise_internet_applications_project/controllers/FileManagementController.java
@@ -1,9 +1,12 @@
 package com.example.enterprise_internet_applications_project.controllers;
 
 import com.example.enterprise_internet_applications_project.models.FileGroup;
+import com.example.enterprise_internet_applications_project.models.Group;
 import com.example.enterprise_internet_applications_project.models.MyFile;
+import com.example.enterprise_internet_applications_project.models.Person;
 import com.example.enterprise_internet_applications_project.services.FileGroupService;
 import com.example.enterprise_internet_applications_project.services.FilesService;
+import com.example.enterprise_internet_applications_project.services.PersonResourceService;
 import com.example.enterprise_internet_applications_project.services.GroupResourceService;
 import com.example.enterprise_internet_applications_project.utils.download.FileDownloadUtil;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,6 +31,9 @@ public class FileManagementController {
 
     @Autowired
     private FilesService storageService;
+
+    @Autowired
+    private PersonResourceService personResourceService;
 
     @Autowired
     private FileGroupService fileGroupService;
@@ -75,7 +81,23 @@ public class FileManagementController {
     }
 
     @PostMapping("/group")
-    public void addFileToGroup(@RequestParam("file_id") Long fileId, @RequestParam("group_id") Long groupId) {
+    public void addFileToGroup(@RequestParam("file_id") Long fileId, @RequestParam("group_id") Long groupId, @RequestParam("userId") Long userId) {
+        Person p = personResourceService.find(userId).get();
+        if (p == null) {
+            throw new IllegalStateException("this user is not belong to this group .. ");
+        }
+
+        boolean ok = false;
+        for (int i = 0; i < p.getPersonGroups().size(); i++) {
+            if(p.getPersonGroups().get(i).getGroup().getId().equals(groupId)){
+                ok = true;
+                break;
+            }
+        }
+        if (!ok) {
+            throw new IllegalStateException("this user is not belong to this group .. ");
+
+        }
         FileGroup fg = new FileGroup(filesService.findFileById(fileId), groupResourceService.findGroupById(groupId));
         for (FileGroup fileGroup : fileGroupService.getAllFileGroups()) {
             if (fileGroup.equals(fg)) {
@@ -83,10 +105,29 @@ public class FileManagementController {
             }
         }
         fileGroupService.addFileToGroup(fileId, groupId);
+
     }
 
     @DeleteMapping("/group")
-    public void removeFileFromGroup(@RequestParam("file_id") Long fileId, @RequestParam("group_id") Long groupId) {
+    public void removeFileFromGroup(@RequestParam("file_id") Long fileId, @RequestParam("group_id") Long groupId, @RequestParam("userId") Long userId) {
+        Person p = personResourceService.find(userId).get();
+        if (p == null) {
+            throw new IllegalStateException("this user is not belong to this group .. ");
+        }
+
+        boolean ok = false;
+        for (int i = 0; i < p.getPersonGroups().size(); i++) {
+            if(p.getPersonGroups().get(i).getGroup().getId().equals(groupId)){
+                ok = true;
+                break;
+            }
+        }
+        if (!ok) {
+            throw new IllegalStateException("this user is not belong to this group .. ");
+        }
+
+
+
         fileGroupService.removeFileFromGroup(fileId, groupId);
     }
 
@@ -101,10 +142,10 @@ public class FileManagementController {
     }
 
     @GetMapping("/getStatusFile")
-    public String statusFile(@RequestParam("nameFile") String nameFile){
-        try{
+    public String statusFile(@RequestParam("nameFile") String nameFile) {
+        try {
             return nameFile + " : " + isCheckIn(nameFile);
-        }catch (Exception e){
+        } catch (Exception e) {
             throw new IllegalStateException("file not found");
         }
     }
@@ -119,9 +160,9 @@ public class FileManagementController {
     }
 
     @GetMapping("/changeStatusFile")
-    public void changeStatusFile(@RequestParam("nameFile") String nameFile, boolean status) throws Exception {
+    public void changeStatusFile(@RequestParam("nameFile") String nameFile, boolean status, Long userId) throws Exception {
         try {
-            storageService.changeStatusFile(status, nameFile);
+            storageService.changeStatusFile(status, nameFile, userId);
         } catch (Exception e) {
             throw new Exception("not find this file : " + nameFile);
         }
@@ -207,7 +248,7 @@ public class FileManagementController {
             if (isCheckIn(nameFile)) {
                 throw new IllegalStateException("You Can't make check in for this file beacuase another user make check in before");
             }
-            changeStatusFile(nameFile, true);
+            changeStatusFile(nameFile, true, 0L);
             downloadFile(id);
         } catch (Exception e) {
             throw new Exception("not find file : " + nameFile);
@@ -215,14 +256,23 @@ public class FileManagementController {
     }
 
     @GetMapping("/check-in")
-    public ResponseEntity<?> checkInFile(@RequestParam("nameFile") String nameFile) throws Exception {
+    public ResponseEntity<?> checkInFile(@RequestParam("nameFile") String nameFile, @RequestParam("userId") Long userId) throws Exception {
         Long id;
         try {
             id = getIdFile(nameFile);
             if (isCheckIn(nameFile) || isPinding(nameFile)) {
                 throw new IllegalStateException("You Can't make check in for this file beacuase another user make check in before");
             }
-            changeStatusFile(nameFile, true);
+
+
+            Long pGId = personResourceService.findGroupIdByUserId(userId);
+            Long fGId = storageService.findGroupIdByFileId(id);
+
+            if (!pGId.equals(fGId)) {
+                throw new IllegalStateException("You Can't make check in for this file beacuase another user make check in before");
+            }
+
+            changeStatusFile(nameFile, true, userId);
             return downloadFile(id);
         } catch (Exception e) {
             throw new Exception("not find file : " + nameFile);
@@ -235,30 +285,35 @@ public class FileManagementController {
     }
 
     @PutMapping("/check-out")
-    public void checkOutFile(@RequestParam("file") MultipartFile file) {
+    public void checkOutFile(@RequestParam("file") MultipartFile file, @RequestParam("userId") Long userId) {
         Long ownerId = ownerIdFile(file.getOriginalFilename());
-        if (findByName(file.getOriginalFilename()) == null) {
+        MyFile myFile = findByName(file.getOriginalFilename());
+        if (myFile == null) {
             uploadFile(file, ownerId);
         }
+
+        if (!myFile.getCheckInUserId().equals(userId)) {
+            throw new IllegalStateException("can not check out using different user id");
+        }
         try {
-            changeStatusFile(file.getOriginalFilename(), false);
+            changeStatusFile(file.getOriginalFilename(), false, 0L);
         } catch (Exception e) {
             throw new IllegalStateException("\"can't change status file to check out\"");
         }
     }
 
     @GetMapping("/readFile")
-    public void readFile(String nameFile){
-        try{
+    public void readFile(String nameFile) {
+        try {
             Long idFile = getIdFile(nameFile);
             downloadFile(idFile);
-        }catch (Exception e){
+        } catch (Exception e) {
             throw new IllegalStateException("can't get id file");
         }
     }
 
     @GetMapping("/files/in/group/{group_id}")
-    public List<String> getFilesInGroup(@PathVariable("group_id") Long groupId){
+    public List<String> getFilesInGroup(@PathVariable("group_id") Long groupId) {
         return fileGroupService.getFilesInGroup(groupId);
     }
 }
